+++ CHANGES.md
@@ -1,15 +1,19 @@
 ## Release notes
 
-### Version v4.3 (2023020500)
+### Version v4.2
+* Add settings option to force date deletion.
+* Add settings option to activate email overriding.
+
+### Version v4.1.3
 * reduce complexity
 * make sure the companion attributes are set correct.
 
-### Version v4.2 (2023020200)
+### Version v4.1.2
 * #1 remove abstraction class with moodle globals
 * #2 optimized privacy api and added privacy test
 * #3 fixed user_login() function
 * #4 companion login clears $_SESSION vars
 
-### Version v4.1 (2022120700)
+### Version v4.1
 * First feature ready version
 * Including phpunit test and behat